+++ deploy/011-deploy-markets.ts
@@ -1,10 +1,11 @@
 import { BigNumber, BigNumberish } from "ethers";
 import { parseUnits } from "ethers/lib/utils";
 import { ethers } from "hardhat";
+import { network } from "hardhat";
 import { DeployFunction, DeployResult } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { InterestRateModels, getConfig, getTokenConfig } from "../helpers/deploymentConfig";
+import { InterestRateModels, getConfig, getTokenConfig, skipRemoteNetworks } from "../helpers/deploymentConfig";
 
 const mantissaToBps = (num: BigNumberish) => {
   return BigNumber.from(num).div(parseUnits("1", 14)).toString();
@@ -18,7 +19,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deploy } = deployments;
 
   const { deployer } = await getNamedAccounts();
-  const { tokensConfig, marketsConfig, preconfiguredAddresses } = await getConfig(hre.network.name);
+  const { tokensConfig, marketsConfig } = await getConfig(hre.network.name);
 
   const comptrollerDeployment = await deployments.get("Unitroller");
 
@@ -64,8 +65,9 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     }
 
     const underlyingDecimals = Number(await tokenContract.decimals());
-
-    console.log(`Deploying VBep20 Proxy for ${symbol} with Implementation ${preconfiguredAddresses.VTokenImpl}`);
+    const normalTimelock = await ethers.getContract("NormalTimelock");
+    const vBep20DelegateDeployment = await deploy("VBep20Delegate", { from: deployer });
+    console.log(`Deploying VBep20 Proxy for ${symbol} with Implementation ${vBep20DelegateDeployment.address}`);
 
     await deploy(`${symbol}`, {
       contract: "VBep20Delegator",
@@ -78,8 +80,8 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
         name,
         symbol,
         VTOKEN_DECIMALS,
-        preconfiguredAddresses.NormalTimelock,
-        preconfiguredAddresses.VTokenImpl,
+        network.live ? normalTimelock.address : deployer,
+        vBep20DelegateDeployment.address,
         EMPTY_BYTES_ARRAY,
       ],
       log: true,
@@ -88,10 +90,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 };
 
 func.tags = ["Markets"];
-func.skip = async hre =>
-  hre.network.name === "sepolia" ||
-  hre.network.name === "opbnbtestnet" ||
-  hre.network.name === "opbnbmainnet" ||
-  hre.network.name === "ethereum";
+func.skip = skipRemoteNetworks();
 
 export default func;